1. fix some warning

+++ src/tm_stat.c
@@ -57,7 +57,7 @@ tm_err_t tm_stat(tm_mdlbin_t* b)
     printf("//Note: PARAM is layer param size, include align padding\r\n\r\n");
     printf("Idx\tLayer\t         outshape\tinoft\toutoft\tPARAM\tMEMOUT OPS\n");
     printf("---\tInput    \t%3d,%3d,%3d\t-   \t0    \t0 \t%ld \t0\n",\
-        idim[1],idim[2],idim[3], idim[1]*idim[2]*idim[3]*sizeof(mtype_t));
+        idim[1],idim[2],idim[3], (long int)(idim[1]*idim[2]*idim[3]*sizeof(mtype_t)));
     //      000  Input    -     224,224,3  0x40001234 0x40004000 100000 500000 200000
     //printf("000  Input    -     %3d,%3d,%d  0x%08x   0x%08x     %6d %6d %6d\n",) 
     int sum_param = 0;
@@ -66,7 +66,7 @@ tm_err_t tm_stat(tm_mdlbin_t* b)
     int layer_i;
     for(layer_i = 0; layer_i < b->layer_cnt; layer_i++){
         tml_head_t* h = (tml_head_t*)(layer_body);
-        TM_DBG("body oft = %d\n", (uint32_t)h - (uint32_t)(b));
+        TM_DBG("body oft = %d\n", (uint32_t)((size_t)h - (size_t)(b)));
         TM_DBG("type=%d, is_out=%d, size=%d, in_oft=%d, out_oft=%d, in_dims=[%d,%d,%d,%d], out_dims=[%d,%d,%d,%d], in_s=%.3f, in_zp=%d, out_s=%.3f, out_zp=%d\n",\
                 h->type,h->is_out,h->size,h->in_oft,h->out_oft,\
                 h->in_dims[0],h->in_dims[1],h->in_dims[2],h->in_dims[3],\
@@ -113,7 +113,7 @@ tm_err_t tm_stat(tm_mdlbin_t* b)
             printf("%03d\t%s      \t%3d,%3d,%3d\t%d\t%d\t%d\t%ld\t", layer_i, tml_str_tbl[h->type], \
                 h->out_dims[1], h->out_dims[2], h->out_dims[3], \
                 h->in_oft, h->out_oft, h->size - tml_headsize_tbl[h->type], \
-                memout*sizeof(mtype_t));
+                (long int)(memout*sizeof(mtype_t)));
             printf("%d\r\n", ops);
         } else {
             return TM_ERR_LAYERTYPE;
